Fixed warnings about unknown z specifier in printf's

+++ src/Util/MemTrack.cpp
@@ -428,7 +428,7 @@ namespace MemTrack
             size_t size = pBlockHeader->GetRequestedSize();
             char const *fileName = pBlockHeader->GetFilename();
             int lineNum = pBlockHeader->GetLineNum();
-            printf("*** #%-6zd %5zd bytes %-50s\n", i, size, typeName);
+            printf("*** #%-6lu %5lu bytes %-50s\n", static_cast<unsigned long>(i), static_cast<unsigned long>(size), typeName);
             printf("... %s:%d\n", fileName, lineNum);
         }
 
@@ -556,16 +556,16 @@ namespace MemTrack
             double totalSizePct = 100.0 * totalSize / grandTotalSize;
 
             printf(
-                "%-50s %5zd %5.1f%% %7zd %5.1f%%\n",
+                "%-50s %5lu %5.1f%% %7lu %5.1f%%\n",
                 pMD->typeName,
-                blockCount,
+                static_cast<unsigned long>(blockCount),
                 blockCountPct,
-                totalSize,
+                static_cast<unsigned long>(totalSize),
                 totalSizePct
             );
         }
         printf("%-50s %5s %5s  %7s %s \n", "--------", "-----", "", "-------", "");
-        printf("%-50s %5zd %5s  %7zd %s \n", "[totals]", grandTotalNumBlocks, "", grandTotalSize, "");
+        printf("%-50s %5lu %5s  %7lu %s \n", "[totals]", static_cast<unsigned long>(grandTotalNumBlocks), "", static_cast<unsigned long>(grandTotalSize), "");
 
         // Clean up.
         free(ppBlockHeader);
